Moved helper function.
