+++ lib/screens/v2/widget_instance/departures.ex
@@ -4,6 +4,7 @@ defmodule Screens.V2.WidgetInstance.Departures do
   alias Screens.Alerts.Alert
   alias Screens.Config.Screen
   alias Screens.Config.V2.FreeTextLine
+  alias Screens.Routes.Route
   alias Screens.Util
   alias Screens.V2.Departure
   alias Screens.V2.WidgetInstance.Departures
@@ -87,7 +88,7 @@ defmodule Screens.V2.WidgetInstance.Departures do
         %{type: :rail} -> :cr
         %{short_name: "SL" <> _} -> :silver
         %{type: :bus} -> :bus
-        %{id: id} -> Util.get_color_for_route(id)
+        %{id: id} -> Route.get_color_for_route(id)
         _ -> ""
       end
 
@@ -104,7 +105,7 @@ defmodule Screens.V2.WidgetInstance.Departures do
         _screen,
         is_only_section
       ) do
-    pill_color = Util.get_color_for_route(route)
+    pill_color = Route.get_color_for_route(route)
 
     text =
       if is_only_section do
