+++ lib/screens/v2/widget_instance/serializer/route_pill.ex
@@ -83,7 +83,7 @@ defmodule Screens.V2.WidgetInstance.Serializer.RoutePill do
           do_serialize(route_id, %{route_name: route_name, gl_branch: true})
       end
 
-    Map.put(route, :color, Util.get_color_for_route(route_id, route_type))
+    Map.put(route, :color, Route.get_color_for_route(route_id, route_type))
   end
 
   @spec serialize_for_audio_departure(Route.id(), String.t(), RouteType.t(), pos_integer() | nil) ::
@@ -129,7 +129,7 @@ defmodule Screens.V2.WidgetInstance.Serializer.RoutePill do
   def serialize_route_for_alert(route_id, gl_long \\ true) do
     route = do_serialize(route_id, %{gl_long: gl_long, gl_branch: true, cr_abbrev: true})
 
-    Map.merge(route, %{color: Util.get_color_for_route(route_id)})
+    Map.merge(route, %{color: Route.get_color_for_route(route_id)})
   end
 
   def serialize_route_for_reconstructed_alert(route_id_group, opts \\ %{})
@@ -146,7 +146,7 @@ defmodule Screens.V2.WidgetInstance.Serializer.RoutePill do
 
   def serialize_route_for_reconstructed_alert({route_id, _}, opts) do
     route = do_serialize(route_id, opts)
-    Map.merge(route, %{color: Util.get_color_for_route(route_id)})
+    Map.merge(route, %{color: Route.get_color_for_route(route_id)})
   end
 
   @typep serialize_opts :: %{
